+++ src/functions.php
@@ -1,8 +1,25 @@
 <?php
 namespace andrefelipe\Orchestrate;
 
+use GuzzleHttp\Client as GuzzleClient;
+
 const DEFAULT_HOST = 'https://api.orchestrate.io';
 const DEFAULT_VERSION = 'v0';
+
+/**
+ * Creates a pre-configured Guzzle Client with the default settings.
+ *
+ * @param string $apiKey  Orchestrate API key. Defaults to getenv('ORCHESTRATE_API_KEY').
+ * @param string $host    Orchestrate API host. Defaults to 'https://api.orchestrate.io'
+ * @param string $version Orchestrate API version. Defaults to 'v0'
+ *
+ * @return \GuzzleHttp\Client
+ */
+function default_http_client($apiKey = null, $host = null, $version = null)
+{
+    $config = default_http_config($apiKey, $host, $version);
+    return new GuzzleClient($config);
+}
 
 /**
  * Form default configuration settings for Guzzle Client.
@@ -10,6 +27,8 @@ const DEFAULT_VERSION = 'v0';
  * @param string $apiKey  Orchestrate API key. Defaults to getenv('ORCHESTRATE_API_KEY').
  * @param string $host    Orchestrate API host. Defaults to 'https://api.orchestrate.io'
  * @param string $version Orchestrate API version. Defaults to 'v0'
+ *
+ * @return array
  */
 function default_http_config($apiKey = null, $host = null, $version = null)
 {