+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## HEAD
+- Added 'default_http_client' function to simplify the Guzzle client instantiation.
+
 ## 0.15.0 - 2015-12-19
 - Introduced the Application object to provide Root Search.
 - Added support for Relation object in searches.